refactor: move OSS survey helpers into onboarding utils
