+++ packages/console/src/pages/OssOnboarding/report-oss-survey.ts
@@ -2,9 +2,7 @@ import { useMemo } from 'react';
 
 import useApi from '@/hooks/use-api';
 
-import { createOssSurveyReporter } from './report-oss-survey.utils';
-
-export type { OssSurveyReportPayload } from './report-oss-survey.utils';
+import { createOssSurveyReporter } from './utils';
 
 const useReportOssSurvey = () => {
   const api = useApi({ hideErrorToast: true });
