+++ packages/console/src/pages/OssOnboarding/utils.ts
@@ -1,6 +1,11 @@
-import { type CompanySize, Project } from '@logto/schemas';
+import { type CompanySize, type OssQuestionnaire, Project } from '@logto/schemas';
+import type { KyInstance } from 'node_modules/ky/distribution/types/ky';
 
-import type { OssSurveyReportPayload } from './report-oss-survey';
+type RequiredOssSurveyField = 'emailAddress' | 'project';
+
+export type OssSurveyReportPayload = Omit<OssQuestionnaire, RequiredOssSurveyField> & {
+  [Key in RequiredOssSurveyField]-?: NonNullable<OssQuestionnaire[Key]>;
+};
 
 export type OssOnboardingFormData = {
   emailAddress: string;
@@ -19,6 +24,20 @@ export const getOssOnboardingDefaultValues = (): OssOnboardingFormData => ({
 });
 
 export const shouldRequireCompanyFields = (project: Project) => project === Project.Company;
+
+export const createOssSurveyReporter = (api: Pick<KyInstance, 'post'>) => {
+  return (payload: OssSurveyReportPayload): void => {
+    void (async () => {
+      try {
+        await api.post('api/oss-survey/report', {
+          keepalive: true,
+          json: payload,
+          retry: { limit: 0 },
+        });
+      } catch {}
+    })();
+  };
+};
 
 export const getOssOnboardingSubmitPayload = (
   data: OssOnboardingFormData
